LR(18) -> LRT
* Add pool for Boss Essence to the menu and spoiler/tracker logs

* Randomize essence drops from Dream Warriors

There are still some glitches with the drop replacement;
particularly, the flashing effect on the Knight persists
until room exit, and the orb counter still appears despite
no Essence being given. The item is, however, correctly
granted, and boss essence drops found elsewhere also appear
to work.

* Make Grimmkin Flames and Boss Essence settings affect early geo

* Remove essence counter HUD and persistent blink effect after killing a boss

Also renamed boss items to reflect the fact that the bosses themselves
are not randomized, but rather their essence.

* Randomize Essence from all Dream Bosses other than Grey Prince Zote

* Correct comment on ReplaceReward

* Randomize Essence from Grey Prince Zote

* Move randomization menu options up to properly accomodate boss essence option

* Put Failed Champion in logic with Wings and Infection

* Add a preview for Grey Prince Zote's item

* Document addition of Boss Essence in README

* Use correct pronoun for the Knight in GPZ hint

(matching how she refers to them in her 3rd diary entry)

* Fix logic for Failed Champion and GPZ

Crossroads_10 (the False Knight room) is not infected in room rando
to prevent the left transition from being blocked, so the infection pogo
with wings is not possible there.

The entrance to Grey Prince Zote only appears if the player has Monarch
Wings. This allows us to simplify the GPZ logic, as alternative ways of
rescuing Bretta without Wings are no longer relevant.

Co-authored-by: Diogo Pinela <diogoid7400@gmail.com>